Update: Add required dependencies for project
The whole point of this is to use TDD so I needed to add and configure
jest as well as make sure that the overall structure of my src makes
at least a little bit of sense. I also needed to use babel to get ES6
modules to play nicely with jest.