Update: remove test module, implement Ship class
No longer need to test that everything is playing nicely, so that stuff
is gone. I added the ship class as per TOP instructions. I was thinking
of doing this in a slightly different way when I initally saw that we
were making battleship but I am happy to follow their guidelines.